Moved component renderers around
