+++ src/graphics/circuitBoard/components/renderer/node/gateRenderer.cpp
