+++ src/graphics/circuitBoard/components/renderer/node/gateRenderer.h
