+++ src/graphics/circuitBoard/components/renderer/node/notNodeRenderer.h
